backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
bcrypt==4.2.1
certifi==2024.12.14
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.17
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pillow==11.0.0
psycopg==3.2.3
psycopg2-binary==2.9.9
PyJWT==2.10.1
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
simplejson==3.19.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.15
